fix(system): curator min_idle_hours Double parse + single-flight backup download

+++ app/src/main/java/com/m57/hermescontrol/data/model/CuratorResponse.kt
@@ -7,7 +7,9 @@ data class CuratorResponse(
     val paused: Boolean? = null,
     val interval_hours: Int? = null,
     val last_run_at: String? = null,
-    val min_idle_hours: Int? = null,
+    // Backend sends a float (e.g. 2.0) — Int here broke JSON parsing (logcat
+    // "Unexpected symbol '.' in numeric literal at path: $.min_idle_hours").
+    val min_idle_hours: Double? = null,
     val stale_after_days: Int? = null,
     val archive_after_days: Int? = null,
 )
